+++ tests/test_cns_schema.py
@@ -161,7 +161,7 @@ class CoreTestCase(unittest.TestCase):
         item_link['out'] = cns_item_out["@id"]
 
         primary_keys = [ item_in['name'], item_out["name"] ]
-        cns_item_link = cnsSchema.cnsConvert(item_link, types, primary_keys, report, primary_keys_lambda=lambda_key_csn_link)
+        cns_item_link = cnsSchema.cnsConvert(item_link, types, primary_keys, report, primary_keys_lambda=lambda_key_cns_link)
         entities.append( cns_item_link )
 
         logging.info(json4debug(entities))